refactoring, remove Hello World
