+++ src/deepresearcher2/__init__.py
@@ -3,12 +3,10 @@ deepresearcher2 public interface
 """
 
 from .logger import logger  # noqa: I001
-from .helloworld import HelloWorld
 from .examples import basic_chat, chat_with_python, mcp_server
 
 __all__ = [
     "logger",
-    "HelloWorld",
     "basic_chat",
     "chat_with_python",
     "mcp_server",
